Move examples/peps/utils.jl code into package
