+++ Project.toml
@@ -12,6 +12,7 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MultiDimDictionaries = "87ff4268-a46e-478f-b30a-76b83dd64e3c"
 NamedGraphs = "678767b0-92e7-4007-89e4-4527a8725b19"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
+SplitApplyCombine = "03a91e81-4c3e-53e1-a0a4-9c0c8f19dd66"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
 
 [compat]
