+++ src/itensornetwork.jl
@@ -94,3 +94,15 @@ function ITensorNetwork(is::IndsNetwork, initstate::Function)
   ψ = insert_links(ψ, edges(is))
   return ψ
 end
+
+# TODO: generalize to other random number distributions
+function randomITensorNetwork(s; link_space)
+  ψ = ITensorNetwork(s; link_space)
+  for v in vertices(ψ)
+    ψᵥ = copy(ψ[v])
+    randn!(ψᵥ)
+    ψᵥ ./= norm(ψᵥ)
+    ψ[v] = ψᵥ
+  end
+  return ψ
+end
