+++ src/utils.jl
@@ -1,0 +1,13 @@
+to_tuple(x) = (x,)
+to_tuple(x::Tuple) = x
+
+function cartesian_to_linear(dims::Tuple)
+  return Dictionary(vec(Tuple.(CartesianIndices(dims))), 1:prod(dims))
+end
+
+# Convert to real if possible
+maybe_real(x::Real) = x
+maybe_real(x::Complex) = iszero(imag(x)) ? real(x) : x
+
+maybe_only(x) = x
+maybe_only(x::Tuple{T}) where {T} = only(x)
